bf(UTAPI-54): create service user using prefix during test setup
