+++ libV2/config/defaults.json
@@ -46,7 +46,7 @@
         "expirationEnabled": false
     },
     "serviceUser": {
-        "arn": "arn:aws:iam::000000000000:user/service-utapi-user",
+        "arn": "arn:aws:iam::000000000000:user/scality-internal/service-utapi-user",
         "enabled": false
     },
     "filter": {
