+++ tests/functional/v2/testEnsureServiceUser.js
@@ -1,11 +1,6 @@
-/* eslint-disable no-console */
 const assert = require('assert');
-const { exec } = require('child_process');
-const path = require('path');
 
 const vaultclient = require('../../utils/vaultclient');
-
-const ensureServiceUser = path.resolve(__dirname, '../../../bin/ensureServiceUser');
 
 const expectedPolicyDocument = {
     Version: '2012-10-17',
@@ -16,42 +11,6 @@ const expectedPolicyDocument = {
     },
 };
 
-async function execPath(path, args, env) {
-    const proc = exec(`${path} ${args.join(' ')}`, {
-        env,
-        stdio: 'pipe',
-    });
-    proc.stdout.on('data', data => console.log(data.toString()));
-    proc.stderr.on('data', data => console.error(data.toString()));
-    return new Promise((resolve, reject) => {
-        proc.on('error', err => reject(err));
-        proc.on('exit', exitCode => {
-            if (exitCode !== 0) {
-                reject(new Error(`ensureServiceUser exited with non-zero code ${exitCode}`));
-                return;
-            }
-            resolve();
-        });
-    });
-}
-
-// Allow overriding the path to the node binary
-// useful to work around issues when running locally and using a node version manager
-const NODE_INTERPRETER = process.env.NODE_INTERPRETER ? process.env.NODE_INTERPRETER : 'node';
-
-function executeScript(account) {
-    return execPath(
-        NODE_INTERPRETER,
-        [ensureServiceUser, 'apply', 'service-utapi-user'],
-        {
-            AWS_ACCESS_KEY_ID: account.accessKey,
-            AWS_SECRET_ACCESS_KEY: account.secretKey,
-            AWS_REGION: 'us-east-1',
-            NODE_TLS_REJECT_UNAUTHORIZED: '0',
-        },
-    );
-}
-
 describe('test bin/ensureServiceUser', () => {
     let adminAccount;
 
@@ -61,7 +20,7 @@ describe('test bin/ensureServiceUser', () => {
 
     after(() => vaultclient.cleanupAccountAndUsers(adminAccount));
 
-    beforeEach(() => executeScript(adminAccount));
+    beforeEach(() => vaultclient.ensureServiceUser(adminAccount));
     afterEach(() => vaultclient.cleanupUsers(adminAccount));
 
     it('should create the service user and attach a policy', async () => {
@@ -74,7 +33,7 @@ describe('test bin/ensureServiceUser', () => {
         const res = await vaultclient.getInternalServiceUserAndPolicies(adminAccount);
         assert.strictEqual(res.policies.length, 1);
         assert.deepStrictEqual(res.policies[0].document, expectedPolicyDocument);
-        await executeScript(adminAccount);
+        await vaultclient.ensureServiceUser(adminAccount);
         const res2 = await vaultclient.getInternalServiceUserAndPolicies(adminAccount);
         assert.strictEqual(res2.policies.length, 1);
         assert.deepStrictEqual(res2.policies[0].document, expectedPolicyDocument);
@@ -85,7 +44,7 @@ describe('test bin/ensureServiceUser', () => {
         assert.strictEqual(detached.length, 1);
         const client = vaultclient.getIAMClient(adminAccount);
         await Promise.all(detached.map(PolicyArn => client.deletePolicy({ PolicyArn }).promise()));
-        await executeScript(adminAccount);
+        await vaultclient.ensureServiceUser(adminAccount);
 
         const res = await vaultclient.getInternalServiceUserAndPolicies(adminAccount);
         assert.strictEqual(res.policies.length, 1);
@@ -94,7 +53,7 @@ describe('test bin/ensureServiceUser', () => {
 
     it('should not create the policy if it already exists', async () => {
         await vaultclient.detachUserPolicies(adminAccount, { name: 'service-utapi-user' });
-        await executeScript(adminAccount);
+        await vaultclient.ensureServiceUser(adminAccount);
         const res = await vaultclient.getInternalServiceUserAndPolicies(adminAccount);
         assert.strictEqual(res.policies.length, 1);
         assert.deepStrictEqual(res.policies[0].document, expectedPolicyDocument);
